+++ RcppTskit/inst/include/RcppTskit_public.hpp
@@ -9,6 +9,7 @@
 // RcppTskit.cpp
 Rcpp::IntegerVector kastore_version();
 Rcpp::IntegerVector tskit_version();
+int rtsk_const_tsk_no_check_integrity();
 
 // sync default options with .cpp!
 SEXP rtsk_treeseq_load(std::string &filename, int options = 0);
@@ -30,6 +31,7 @@ SEXP rtsk_treeseq_get_num_populations(SEXP ts);
 SEXP rtsk_treeseq_get_num_migrations(SEXP ts);
 SEXP rtsk_treeseq_get_num_individuals(SEXP ts);
 SEXP rtsk_treeseq_get_num_samples(SEXP ts);
+Rcpp::IntegerVector rtsk_treeseq_get_samples(SEXP ts);
 SEXP rtsk_treeseq_get_num_nodes(SEXP ts);
 SEXP rtsk_treeseq_get_num_edges(SEXP ts);
 SEXP rtsk_treeseq_get_num_trees(SEXP ts);
@@ -61,6 +63,7 @@ Rcpp::String rtsk_table_collection_get_file_uuid(SEXP tc);
 bool rtsk_table_collection_has_index(SEXP tc, int options = 0);
 void rtsk_table_collection_build_index(SEXP tc, int options = 0);
 void rtsk_table_collection_drop_index(SEXP tc, int options = 0);
+void rtsk_table_collection_sort(SEXP tc, int edge_start = 0, int options = 0);
 Rcpp::List rtsk_table_collection_summary(SEXP tc);
 Rcpp::List rtsk_table_collection_metadata_length(SEXP tc);
 int rtsk_individual_table_add_row(
@@ -71,6 +74,7 @@ int rtsk_individual_table_add_row(
 int rtsk_node_table_add_row(
     SEXP tc, int flags = 0, double time = 0, int population = -1,
     int individual = -1, Rcpp::Nullable<Rcpp::RawVector> metadata = R_NilValue);
+Rcpp::List rtsk_node_table_get_row(SEXP tc, int row_id);
 int rtsk_edge_table_add_row(
     SEXP tc, double left, double right, int parent, int child,
     Rcpp::Nullable<Rcpp::RawVector> metadata = R_NilValue);